fixup! refactor(cloud-function): rename {plans => stripePlans}
